% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSmix_functions_package.R
\name{print.bootMSmix}
\alias{print.bootMSmix}
\title{Print of the bootstrap confidence intervals for the fitted mixture of Mallows models with Spearman distance}
\usage{
\method{print}{bootMSmix}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"bootMSmix"} returned by \code{\link{bootstrapMSmix}}.}

\item{...}{Further arguments passed to or from other methods (not used).}
}
\description{
\code{print} method for class \code{"bootMSmix"}.
}
