% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSmix_functions_package.R
\name{print.bootMSmoe}
\alias{print.bootMSmoe}
\title{Print of the bootstrap confidence intervals for mixtures of experts of
Mallows models with Spearman distance}
\usage{
\method{print}{bootMSmoe}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"bootMSmoe"} returned by \code{\link{bootstrapMSmoe}}.}

\item{...}{Further arguments passed to or from other methods (not used).}
}
\description{
\code{print} method for class \code{"bootMSmoe"}.
}
