% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSmix_functions_package.R
\name{bootstrapMSmoe}
\alias{bootstrapMSmoe}
\alias{plot.bootMSmoe}
\title{Bootstrap confidence intervals for mixture of experts of Mallows models with Spearman distance}
\usage{
bootstrapMSmoe(
  object,
  n_boot = 50,
  type = "soft",
  conf_level = 0.95,
  all = FALSE,
  n_start = 10,
  parallel = FALSE
)

\method{plot}{bootMSmoe}(x, ...)
}
\arguments{
\item{object}{An object of class \code{"emMSmoe"} returned by \code{\link{fitMSmoe}}.}

\item{n_boot}{Number of desired bootstrap samples. Defaults to 50.}

\item{type}{Character indicating which bootstrap method must be used. Available options are: \code{"soft"} or \code{"separated"}. Defaults to \code{"soft"}. See Details.}

\item{conf_level}{Value in the interval (0,1] indicating the desired confidence level of the interval estimates. Defaults to 0.95.}

\item{all}{Logical: whether the bootstrap samples of the MLEs for all the parameters must be returned. Defaults to \code{FALSE}.}

\item{n_start}{Number of starting points for the MLE on each bootstrap sample. Defaults to 10.}

\item{parallel}{Logical: whether parallelization over multiple initializations of the EM algorithm must be used. Used when \code{rankings} contains some partial rankings. Defaults to \code{FALSE}.}

\item{x}{An object of class \code{"bootMSmoe"} returned by \code{\link{bootstrapMSmoe}}.}

\item{...}{Further arguments passed to or from other methods (not used).}
}
\value{
An object of class \code{"bootMSmoe"}, namely a list with the following named components:
\describe{
\item{\code{itemwise_ci_rho}}{The bootstrap itemwise confidence intervals for the component-specific consensus rankings.}
\item{\code{ci_boot_theta}}{The bootstrap confidence intervals for the component-specific precisions.}
\item{\code{ci_boot_betas}}{The bootstrap confidence intervals for the GLM parameters. Returned when \code{type = "soft"}, otherwise \code{NULL}.}
\item{\code{boot}}{List containing all the \code{n_boot} bootstrap parameters. Returned when \code{all = TRUE}, otherwise \code{NULL}.}
}

The \code{boot} sublist contains the following named components:
\describe{
\item{\code{rho_boot}}{List of length \code{n_clust} with the bootstrap MLEs of the consensus rankings. Each element of the list is an integer \code{n_boot} \eqn{\times}{x} \code{n_items} matrix containing, in each row, the bootstrap MLEs of the consensus ranking for a specific component.}
\item{\code{theta_boot}}{Numeric \code{n_boot}\eqn{\times}{x} \code{n_clust} matrix with the bootstrap MLEs of the component-specific precision parameters in each row.}
\item{\code{beta_boot}}{Numeric \code{n_clust-1}\eqn{\times}{x} \code{H+1}\eqn{\times}{x} \code{n_boot} array with the bootstrap GLM parameters in each slice. Returned when\code{type = "soft"}, otherwise \code{NULL}.}
}

For the component-specific bootstrap consensus ranking estimates, a heatmap is returned.

For the component-specific precisions a kernel density plot is returned.

For the GLM parameters a box-plot is returned.
}
\description{
Return the bootstrap confidence intervals for the parameters of a mixture of experts of Mallow models with Spearman distance fitted on full or partial rankings.

\code{plot} method for class \code{"bootMSmoe"}.
}
\details{
Two types of bootstrap are available: 1) \code{type = "soft"} (default), which is
the soft-separated bootstrap (Crispino et al., 2025) and returns confidence intervals for all
the parameters of the mixture of experts of Mallow models with Spearman distance; 2) \code{type = "separated"}, which is the separated bootstrap
(Taushanov and Berchtold, 2019) and returns bootstrap samples only for the component-specific
consensus rankings and precisions.
}
\references{
Crispino M, Mollica C and Modugno L (2025). MSmix: An R Package for clustering partial rankings via mixtures of Mallows Models with Spearman distance. \emph{(submitted)}

Taushanov Z and Berchtold A (2019). Bootstrap validation of the estimated parameters in mixture models used for clustering. \emph{Journal de la société française de statistique}, \bold{160}(1).

Efron B (1982). The Jackknife, the Bootstrap, and Other Resampling Plans. Philadelphia, \emph{Pa. :Society for Industrial and Applied Mathematics}.
}
