% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSmix_functions_package.R
\name{rearrange_output_mix}
\alias{rearrange_output_mix}
\title{Arrange the output of MLE of mixtures of Mallows models with Spearman distance via EM algorithms}
\usage{
rearrange_output_mix(output, ord)
}
\arguments{
\item{output}{An object of class \code{"emMSmix"} returned by \code{\link{fitMSmix}}.}

\item{ord}{Integer vector of length \code{n_clust} with the desired relabelling of the mixture component labels.}
}
\description{
Arrange the output of the object of class \code{"emMSmix"} according to a given relabelling of the mixture component labels.
}
