% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSmix_functions_package.R
\name{rearrange_output_moe}
\alias{rearrange_output_moe}
\title{Arrange the output of MLE of MoE of Mallows models with Spearman distance via EM algorithms}
\usage{
rearrange_output_moe(output, ord)
}
\arguments{
\item{output}{An object of class \code{"emMSmoe"} returned by \code{\link{fitMSmoe}}.}

\item{ord}{Integer vector of length \code{n_clust} with the desired relabelling of the mixture component labels.}
}
\description{
Arrange the output of the object of class \code{"emMSmoe"} according to a given relabelling of the mixture component labels.
}
