% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSmix_functions_package.R
\name{seMSmoe}
\alias{seMSmoe}
\title{Standard errors for mixtures of Mallows models with Spearman distance}
\usage{
seMSmoe(object)
}
\arguments{
\item{object}{An object of class \code{"emMSmoe"} returned by \code{\link{fitMSmoe}}.}
}
\value{
A list with the following named components:

\item{\code{ci_theta}}{The confidence intervals for the precision parameters.}
\item{\code{ci_weights}}{The confidence intervals for the mixture weights (when \eqn{G>1}), otherwise \code{NULL}.}
}
\description{
Return the (Hessian-based) standard errors of the continuous parameters of a mixture of Mallows models with Spearman distance fitted to full rankings, namely the component-specific precisions and weights.
}
\details{
The current implementation of the standard errors assumes that the observed rankings are complete.
}
\references{
Crispino M, Mollica C and Modugno L (2025). MSmix: An R Package for clustering partial rankings via mixtures of Mallows Models with Spearman distance. \emph{(submitted)}

Marden JI (1995). Analyzing and modeling rank data. \emph{Monographs on Statistics and Applied Probability} (64). Chapman & Hall, ISSN: 0-412-99521-2. London.

McLachlan G and Peel D (2000). Finite Mixture Models. \emph{Wiley Series in Probability and Statistics}, John Wiley & Sons.
}
