% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSmix_functions_package.R
\name{summary.emMSmoe}
\alias{summary.emMSmoe}
\alias{print.summary.emMSmoe}
\title{Summary of the fitted mixture of experts of Mallows models with Spearman distance}
\usage{
\method{summary}{emMSmoe}(object, digits = 3, ...)

\method{print}{summary.emMSmoe}(x, ...)
}
\arguments{
\item{object}{An object of class \code{"emMSmoe"} returned by \code{\link{fitMSmoe}}.}

\item{digits}{Integer: decimal places for rounding the numerical summaries. Defaults to 3.}

\item{...}{Further arguments passed to or from other methods (not used).}

\item{x}{An object of class \code{"summary.emMSmoe"} returned by \code{\link{summary.emMSmoe}}.}
}
\value{
A list with the following named components:

\item{\code{modal_rankings}}{Integer matrix with the MLEs of the \eqn{G} component-specific consensus rankings in each row.}
\item{\code{modal_orderings}}{Character matrix with the MLEs of the \eqn{G} component-specific consensus orderings in each row.}
\item{\code{theta}}{Numeric vector of the MLEs of the \eqn{G} precisions.}
\item{\code{betas}}{Numeric matrix of the MLEs of the GLM parameters.}
\item{\code{MAP_distr}}{Percentage distribution of the component memberships based on the MAP allocation. Returned when \code{n_clust > 1}, otherwise \code{NULL}}
\item{\code{conv_perc}}{Percentage of convergence of the EM algorithm over the multiple starting points.}
\item{\code{BIC}}{BIC value (based on full or augmented rankings).}
\item{\code{BIC_part}}{BIC value (based on partial rankings).}
\item{\code{call}}{The matched call.}
}
\description{
\code{summary} method for class \code{"emMSmoe"}.

\code{print} method for class \code{"summary.emMSmoe"}.
}
\seealso{
\code{\link{fitMSmoe}}, \code{\link{plot.emMSmoe}}
}
